refactor: consolidate platform code into src/platform/ as single authority
Reorganize the platform abstraction layer so src/platform/ owns all
definitions. No re-export chains — consumers import directly:

  Interfaces:  @platform/interfaces/{config,state,log,filesystem,...}
  Defaults:    @platform/defaults/{nodeFilesystem,consoleLog,memoryState,...}
  Context:     @platform/platform (initPlatform, platform, tryPlatform)

src/agent/core/ facades become pure thin wrappers (6-16 lines each)
that delegate to platform(). Interface definitions, Node.js defaults,
and the console log backend are moved from src/agent/core/ to their
canonical locations in src/platform/.

Also removes the unused nodeFileSystem export (Bugbot feedback).
