+++ src/agent/core/config.ts
@@ -1,17 +1,15 @@
 /**
- * Platform-agnostic configuration facade for the agent core.
+ * Configuration facade — convenience wrapper over platform().config.
  *
- * Thin wrapper over `platform().config`. Consumer code imports this
- * module for convenience; the canonical definition lives in
- * `@platform/interfaces`.
+ * Falls back to defaultValue before platform initialization to
+ * support module-level calls.
  */
 import { tryPlatform } from '@platform/platform';
+import type { ConfigProvider } from '@platform/interfaces/config';
 
-export interface ConfigProvider {
-  get<T>(key: string, defaultValue?: T): T;
-}
+// Re-export the type so existing `import { ConfigProvider } from '@agent/core/config'` works.
+export type { ConfigProvider };
 
-/** Read a configuration value. Returns defaultValue if platform not yet initialized. */
 export function getConfig<T>(key: string, defaultValue?: T): T {
   const p = tryPlatform();
   return p ? p.config.get(key, defaultValue) : (defaultValue as T);
