+++ src/frontend/vscode/vscodeSecrets.ts
@@ -8,6 +8,7 @@ import * as vscode from 'vscode';
 
 import type { PlatformSecrets } from '@platform/secrets';
 
+
 export class VscodeSecrets implements PlatformSecrets {
   private readonly storage: vscode.SecretStorage;
