+++ src/platform/interfaces/config.ts
@@ -1,0 +1,6 @@
+/**
+ * Platform configuration provider interface.
+ */
+export interface ConfigProvider {
+  get<T>(key: string, defaultValue?: T): T;
+}
